+++ tests/core/tests/test_resources.py
@@ -477,6 +477,8 @@ class ModelResourceTest(TestCase):
         self.assertTrue(result.rows[0].diff)
         self.assertEqual(result.rows[0].import_type,
                          results.RowResult.IMPORT_TYPE_UPDATE)
+        self.assertEqual(result.rows[0].row_values.get('name'), None)
+        self.assertEqual(result.rows[0].row_values.get('author_email'), None)
 
         instance = Book.objects.get(pk=self.book.pk)
         self.assertEqual(instance.author_email, 'test@example.com')
@@ -2191,3 +2193,30 @@ class BulkUUIDBookDeleteTest(BulkTest):
         self.resource.import_data(self.dataset)
         self.assertEqual(0, UUIDBook.objects.count())
 
+
+class RawValueTest(TestCase):
+    def setUp(self):
+        class _BookResource(resources.ModelResource):
+            class Meta:
+                model = Book
+                store_row_values = True
+
+        self.resource = _BookResource()
+
+        self.book = Book.objects.create(name="Some book")
+        self.dataset = tablib.Dataset(headers=['id', 'name', 'author_email',
+                                               'price'])
+        row = [self.book.pk, 'Some book', 'test@example.com', "10.25"]
+        self.dataset.append(row)
+
+    def test_import_data(self):
+        result = self.resource.import_data(self.dataset, raise_errors=True)
+
+        self.assertFalse(result.has_errors())
+        self.assertEqual(len(result.rows), 1)
+        self.assertTrue(result.rows[0].diff)
+        self.assertEqual(result.rows[0].import_type,
+                         results.RowResult.IMPORT_TYPE_UPDATE)
+        self.assertEqual(result.rows[0].row_values.get('name'), 'Some book')
+        self.assertEqual(result.rows[0].row_values.get('author_email'), 'test@example.com')
+        self.assertEqual(result.rows[0].row_values.get('price'), '10.25')